cmd/thor: improve error handling
